requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
attrs==22.1.0
awsebcli==3.20.3
bcrypt==4.0.0
beautifulsoup4==4.11.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.3
cryptography==38.0.1
cycler==0.11.0
deta==1.1.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
easyocr==1.4.2
fastapi==0.75.2
fonttools==4.37.3
future==0.16.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
imageio==2.17.0
jmespath==0.10.0
jsonschema==3.2.0
kiwisolver==1.4.4
matplotlib==3.5.1
networkx==2.8
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging==21.3
paramiko==2.11.0
pathspec==0.9.0
Pillow==9.1.0
pycparser==2.21
pydantic==1.9.0
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-multipart==0.0.5
PyWavelets==1.3.0
PyYAML==5.4.1
requests==2.26.0
scikit-image==0.19.2
scipy==1.8.0
semantic-version==2.8.5
six==1.14.0
sniffio==1.2.0
soupsieve==2.3.2.post1
starlette==0.17.1
termcolor==1.1.0
texttable==1.6.4
tifffile==2022.4.22
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.8.1+cpu 
torchvision==0.9.1+cpu
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn[standard]==0.17.6
wcwidth==0.1.9
websocket-client==0.59.0
